pyproject.toml
[tool.poetry]
name = "r2-d2"
version = "0.1.0"
description = ""
authors = ["Xuan <37977109+AprilNEA@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "r2_d2"}]

[tool.poetry.dependencies]
python = "^3.10"
telethon = "^1.26.1"
pyyaml = "^6.0"
aiohttp = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
